Revert move and update some dependencies
