+++ test/src/test/java/org/frankframework/runner/Log4j2LogConsumer.java
